+++ app/src/main/java/io/github/aw1y2z/sesame/data/AppConfig.java
@@ -30,6 +30,7 @@ public class AppConfig {
 
     private Boolean enableForestLog = true;
     private Boolean enableGoldenBeansLog = true;
+    private Boolean enableCaptchaLog = true;
     private Boolean enableFarmLog = true;
     private Boolean enableOtherLog = true;
     private Boolean enableDebugLog = false;
@@ -72,6 +73,8 @@ public class AppConfig {
 
     public Boolean getEnableGoldenBeansLog() { return enableGoldenBeansLog; }
     public void setEnableGoldenBeansLog(Boolean value) { enableGoldenBeansLog = value; }
+    public Boolean getEnableCaptchaLog() { return enableCaptchaLog; }
+    public void setEnableCaptchaLog(Boolean value) { enableCaptchaLog = value; }
 
     public Boolean getEnableFarmLog() { return enableFarmLog; }
     public void setEnableFarmLog(Boolean value) { enableFarmLog = value; }