+++ app/src/main/java/io/github/aw1y2z/sesame/ui/miuix/MiuixLogViewerActivity.kt
@@ -77,6 +77,7 @@ enum class LogType(val displayName: String) {
     GOLDENBEANS("金豆记录"),
     FARM("庄园记录"),
     OTHER("其他记录"),
+    CAPTCHA("验证记录"),
     DEBUG("抓包记录"),
     ERROR("查看异常日志"),
     RUNTIME("查看运行日志");
@@ -88,6 +89,7 @@ enum class LogType(val displayName: String) {
             GOLDENBEANS -> FileUtil.getGoldenBeansLogFile()
             FARM -> FileUtil.getFarmLogFile()
             OTHER -> FileUtil.getOtherLogFile()
+            CAPTCHA -> FileUtil.getCaptchaLogFile()
             DEBUG -> FileUtil.getDebugLogFile()
             ERROR -> FileUtil.getErrorLogFile()
             RUNTIME -> FileUtil.getRuntimeLogFile()